Place unit tests in test case classes
There is enough tests to make using and debugging them unwieldy. Moving
them into separate classes is the first step on the way to making a test
framework in the future. For now this is enough and I'll do more
improvements when it become obvious what I want to improve and how.

Passing tests will print their success to stdout now.