Test runners can find unit tests through reflection
This avoids the issue of needed to specifically call each individual
test method. There is a new issue however in that the tests are not
being run in the expected order. Since the tests share state they will
start in an unexpected state and fail.

Now is the time for isolating test state.